🐛 FIX: Donation/Evolution system calculation errors
## Problem
- Level-up was resetting accumulators to 0, losing excess donations
- Evolution bar showed 0% (Web UI) or 100% (Discord) incorrectly
- Power accumulator was reset instead of carried over
- Type conversion issue: float values truncated to int

## Root Cause
In progress_service.py apply_donation_units():
- snap.evo_acc was reset to 0 instead of carrying over excess
- snap.power_acc was reset to 0 (or 100 for exact hit) instead of keeping accumulated value
In mech_data_store.py BarsCompat:
- mech_progress_current/max were int type, losing decimal precision

## Solution
1. Carry over excess evolution points after level-up
2. Keep accumulated power and only add bonus for exact hit
3. Changed bar fields to float type for decimal precision

## Testing
With $15 donated (3x $5):
- Level: 2 ✅
- Power: $15.00 / $16.00 = 93% ✅
- Evolution: $4.50 / $15.00 = 30% ✅
(Previously showed Power 0%, Evolution 0%/100%)

+++ services/mech/mech_data_store.py
@@ -42,8 +42,8 @@ class BarsCompat:
     """Legacy compatibility object for bars data."""
     Power_current: float = 0.0  # Support decimal power values (e.g. 0.99)
     Power_max_for_level: int = 100
-    mech_progress_current: int = 0
-    mech_progress_max: int = 20
+    mech_progress_current: float = 0.0  # Changed to float for decimal values
+    mech_progress_max: float = 20.0  # Changed to float for decimal values
 
 
 # ============================================================================ #
